requirements.txt
requests
python-dotenv
# pandas
